requirements.txt
pony
watchfiles
discord.py
emoji